第8节完成

+++ src/main/java/com/imooc/service/search/ISearchService.java
@@ -46,7 +46,7 @@ public interface ISearchService {
     /**
      * 聚合特定小区的房间数
      */
-    ServiceResult<Long> aggregateDistrictHouse(String cityEnName, String regionEnName, String district);
+    ServiceResult<Integer> aggregateDistrictHouse(String cityEnName, String regionEnName, String district);
 
     /**
      * 聚合城市数据
@@ -67,4 +67,5 @@ public interface ISearchService {
      * @return
      */
     ServiceMultiResult<Long> mapQuery(MapSearch mapSearch);
+    
 }
